updates percentRead to be a float in graphql and a decimal in postgres

+++ database/tables.go
@@ -85,7 +85,7 @@ var tables = []table{
 			user_id int NOT NULL REFERENCES users(id),
 			content_item_id int NOT NULL REFERENCES content_items(id),
 			read_state varchar NOT NULL,
-			percent_read int,
+			percent_read decimal,
 			created_at timestamp with time zone NOT NULL,
 			updated_at timestamp with time zone NOT NULL,
 			UNIQUE(user_id, content_item_id)
